Setting White Balance and Exposure of MT9D111 manually
AutoWB and AutoExposure are disabled and custom values for related
parameters (ColorCorrectionMatrix, Digital Gain, Analog Gain, shutter
width and shutter delay) are manually fed. The custom values are first
obtained by enabling auto first and collecting the parameter values it
has generated.
Enhancement: We could collect the parameter values in a calibration
phase from the Auto mode in a global variable and then use those values,
instead of looking for the message in UART and then hard-coding the
values.