Functionality complete
CC3200 brought up in AP mode, web page served, Wifi credentials
collected from user and stored in Flash  in ProvisioningAP(). Storage is
in the form of string with fields separated by nulls.
Codes to read file from flash and parse it to obtain the wifi
credentials and use it to connect to WiFi AP also added.
These codes are called before while(1); in Main_Task(); in main.c file